requirements.txt
streamlit==1.24.0
pandas==2.0.3
numpy<2.0.0
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
tensorflow==2.13.0
pandas_datareader==0.10.0
plotly==5.15.0
python-dateutil==2.8.2 